+++ lib/src/core/model/data/environment/cli_tool.dart
@@ -65,6 +65,14 @@ enum CliTool {
     name: 'Sentry CLI',
     commandName: 'sentry-cli',
   ),
+  lcov(
+    name: 'Lcov',
+    commandName: 'lcov',
+  ),
+  java(
+    name: 'Java',
+    commandName: 'java',
+  ),
   brew(
     name: 'brew',
     commandName: 'brew',